Panel dims to uint16: one 256x64 panel is the correct HUB75 chain 2D config
Justin's glass photos decoded the residual scramble: setUpMatrix numbers
pixels panel-major (per-panel consecutive) while the HUB75 bus blits
row-major - every probe element (repeated F per panel, folded green line,
per-panel red dashes, corner dot displacement) matches that mismatch
exactly, and it also explains the historic pre-bypass staircase (the
virtual layer was innocent; host-sim had already proven identity).

The only 2D layout whose ledmap is identity is ONE 256x64 panel - exactly
WLED-MM's documented cure - which uint8_t Panel.width made impossible
(256 wraps to 0). Widened width/height to uint16_t; the one-panel config
now survives the cfg round-trip and the ledmap is identity.

Device reconfigured to mpc1/256x64 and re-verified: text stable at fx 122,
30 fps, heap flat. Recipe reversed in bench file + QA (four-64s layout is
now marked DO NOT USE for HUB75 chains). Upstream ledger: item 8 added,
item 5 rewritten to final form.

+++ wled00/FX.h
@@ -975,8 +975,8 @@ class WS2812FX {
     struct Panel {
       uint16_t xOffset; // x offset relative to the top left of matrix in LEDs
       uint16_t yOffset; // y offset relative to the top left of matrix in LEDs
-      uint8_t  width;   // width of the panel
-      uint8_t  height;  // height of the panel
+      uint16_t width;   // width of the panel; 16-bit so a HUB75 chain can be one wide panel (4x64=256 wraps to 0 in 8 bits)
+      uint16_t height;  // height of the panel
       union {
         uint8_t options;
         struct {
